Renames config/get to config/getConfig Removes CONF_ENV and refactors environment variable resolution to use NODE_ENV again.
