+++ config/getConfig.js
