+++ internal/scripts/build.js
@@ -5,7 +5,7 @@ import appRootDir from 'app-root-dir';
 import { resolve as pathResolve } from 'path';
 import webpackConfigFactory from '../webpack/configFactory';
 import { exec } from '../utils';
-import getConfig from '../../config/get';
+import getConfig from '../../config/getConfig';
 
 // First clear the build output dir.
 exec(`rimraf ${pathResolve(appRootDir.get(), getConfig('buildOutputPath'))}`);
