+++ internal/scripts/clean.js
@@ -3,7 +3,7 @@
 import { resolve as pathResolve } from 'path';
 import appRootDir from 'app-root-dir';
 import { exec } from '../utils';
-import getConfig from '../../config/get';
+import getConfig from '../../config/getConfig';
 
 const cmd = `$(npm bin)/rimraf ${pathResolve(appRootDir.get(), getConfig('buildOutputPath'))}`;
