+++ package.json
@@ -12,10 +12,10 @@
     "build": "cross-env NODE_ENV=production babel-node internal/scripts/build",
     "clean": "babel-node internal/scripts/clean",
     "deploy": "babel-node internal/scripts/deploy",
-    "development": "babel-node internal/development",
+    "development": "cross-env NODE_ENV=development babel-node internal/development",
     "lint": "eslint src config tools",
     "preinstall": "node internal/scripts/preinstall",
-    "start": "node build/server",
+    "start": "cross-env NODE_ENV=production node build/server",
     "test:coverage": "jest --coverage",
     "test": "jest"
   },
@@ -69,19 +69,18 @@
     "compression": "1.6.2",
     "cross-env": "3.1.4",
     "dotenv": "4.0.0",
-    "express": "4.14.0",
+    "express": "4.14.1",
     "helmet": "3.4.0",
     "hpp": "0.2.1",
     "modernizr": "3.3.1",
     "normalize.css": "5.0.0",
-    "offline-plugin": "4.5.5",
+    "offline-plugin": "4.6.1",
     "react": "15.4.2",
-    "react-async-component": "0.2.1",
+    "react-async-component": "0.2.2",
     "react-dom": "15.4.2",
     "react-helmet": "4.0.0",
     "react-router": "4.0.0-alpha.6",
     "serialize-javascript": "1.3.0",
-    "user-home": "2.0.0",
     "uuid": "3.0.1"
   },
   "devDependencies": {
@@ -105,32 +104,32 @@
     "css-loader": "0.26.1",
     "enzyme": "2.7.1",
     "enzyme-to-json": "1.4.5",
-    "eslint": "3.14.0",
-    "eslint-config-airbnb": "14.0.0",
+    "eslint": "3.15.0",
+    "eslint-config-airbnb": "14.1.0",
     "eslint-plugin-flowtype": "2.30.0",
     "eslint-plugin-import": "2.2.0",
-    "eslint-plugin-jsx-a11y": "3.0.2",
+    "eslint-plugin-jsx-a11y": "4.0.0",
     "eslint-plugin-react": "6.9.0",
     "extract-text-webpack-plugin": "2.0.0-beta.5",
-    "file-loader": "0.9.0",
+    "file-loader": "0.10.0",
     "glob": "7.1.1",
     "happypack": "3.0.2",
-    "html-webpack-plugin": "2.26.0",
+    "html-webpack-plugin": "2.28.0",
     "jest": "18.1.0",
     "md5": "2.2.1",
     "modernizr-loader": "1.0.1",
-    "node-notifier": "4.6.1",
+    "node-notifier": "5.0.2",
     "react-addons-test-utils": "15.4.2",
     "react-hot-loader": "3.0.0-beta.6",
     "regenerator-runtime": "0.10.1",
     "rimraf": "2.5.4",
     "semver": "5.3.0",
-    "source-map-support": "0.4.10",
+    "source-map-support": "0.4.11",
     "style-loader": "0.13.1",
-    "webpack": "2.2.0",
+    "webpack": "2.2.1",
     "webpack-bundle-analyzer": "2.2.1",
-    "webpack-dev-middleware": "1.9.0",
-    "webpack-hot-middleware": "2.15.0",
+    "webpack-dev-middleware": "1.10.0",
+    "webpack-hot-middleware": "2.16.1",
     "webpack-md5-hash": "0.0.5",
     "webpack-node-externals": "1.5.4"
   }
